+++ common/src/main/java/earth/terrarium/cadmus/common/compat/prometheus/PrometheusIntegration.java
@@ -1,5 +1,7 @@
 package earth.terrarium.cadmus.common.compat.prometheus;
 
+import com.teamresourceful.resourcefullib.common.utils.TriState;
+import com.teamresourceful.resourcefullib.common.utils.modinfo.ModInfoUtils;
 import earth.terrarium.cadmus.api.claims.maxclaims.MaxClaimProviderApi;
 import earth.terrarium.cadmus.common.compat.prometheus.roles.CadmusOptions;
 import earth.terrarium.cadmus.common.compat.prometheus.roles.client.CadmusOptionsDisplay;
@@ -16,6 +18,10 @@ import java.util.UUID;
 
 public class PrometheusIntegration {
     public static final ResourceLocation PROMETHEUS_ID = new ResourceLocation(Prometheus.MOD_ID, Prometheus.MOD_ID);
+
+    public static boolean prometheusLoaded() {
+        return ModInfoUtils.isModLoaded("prometheus");
+    }
 
     public static void registerClient() {
         var api = PermissionApi.API;
@@ -39,6 +45,12 @@ public class PrometheusIntegration {
         RoleOptionsApi.API.register(CadmusOptions.SERIALIZER);
         MaxClaimProviderApi.API.register(PROMETHEUS_ID, new PrometheusMaxClaimProvider());
         MaxClaimProviderApi.API.setSelected(PROMETHEUS_ID);
+        PermissionApi.API.addDefaultPermission(CadmusAutoCompletes.PERSONAL_BLOCK_BREAKING, TriState.TRUE);
+        PermissionApi.API.addDefaultPermission(CadmusAutoCompletes.PERSONAL_BLOCK_PLACING, TriState.TRUE);
+        PermissionApi.API.addDefaultPermission(CadmusAutoCompletes.PERSONAL_BLOCK_EXPLOSIONS, TriState.TRUE);
+        PermissionApi.API.addDefaultPermission(CadmusAutoCompletes.PERSONAL_BLOCK_INTERACTIONS, TriState.TRUE);
+        PermissionApi.API.addDefaultPermission(CadmusAutoCompletes.PERSONAL_ENTITY_INTERACTIONS, TriState.TRUE);
+        PermissionApi.API.addDefaultPermission(CadmusAutoCompletes.PERSONAL_ENTITY_DAMAGE, TriState.TRUE);
     }
 
     public static boolean hasPermission(Player player, String permission) {