+++ changelog.md
@@ -1,2 +1,4 @@
-- Fixed /claim settings tied to same Prometheus permissions as bypassing claims. #26
-- Fixed being able to unclaim admin chunks with /unclaim command
+- 1.20.4 Port
+- Added better claim settings errors. It'll now tell you if you don't have permission
+on your team or if your role doesn't have the permission.
+- Claim settings permissions are now enabled by default for Prometheus.